chore/updated folder structure, renamed npm script and updated readme
