+++ package.json
@@ -13,7 +13,7 @@
 		"lint:js": "wp-scripts lint-js",
 		"packages-update": "wp-scripts packages-update",
 		"plugin-zip": "wp-scripts plugin-zip",
-		"start": "wp-scripts start"
+		"watch:dev": "wp-scripts start"
 	},
 	"devDependencies": {
 		"@wordpress/scripts": "^27.9.0"
