--- readme.md
@@ -1,11 +1,0 @@
-## Site Settings Plugin
-
-## Installation
-
-- Setup a new multisite using `bbdocker create`
-- Use `multisite.bigbite.site` with subdomain option
-- Clone this repo into `wp-content/plugins/site-settings`
-- `cd wp-content/plugins/site-settings`
-- `npm install`
-- `composer install`
-- `npm start`
