+++ src/admin/schema/supportedCategories.js
