+++ src/admin/styles/style.scss
@@ -1,0 +1,5 @@
+.bb-site-settings {
+	background-color: #21759b;
+	color: #fff;
+	padding: 2px;
+}
